driver: rename and bind to Hermes only
We will extend the XDMA driver to handle Hermes devices only. Hence,
make it bind only to Hermes Vendor/Device ID and rename the module.
