+++ src/driver/10-hermes.rules
@@ -1,0 +1,1 @@
+KERNEL=="hermes*",MODE="0666",SYMLINK="%k"
