--- src/driver/10-xdma.rules
@@ -1,1 +1,0 @@
-KERNEL=="xdma*",MODE="0666",SYMLINK="%k"
