+++ src/driver/hermes_mod.c
@@ -31,17 +31,18 @@
 /* include early, to verify it depends only on the headers above */
 #include "libxdma_api.h"
 #include "libxdma.h"
-#include "xdma_mod.h"
+#include "hermes_mod.h"
 #include "xdma_cdev.h"
 #include "version.h"
 
-#define DRV_MODULE_NAME		"xdma"
-#define DRV_MODULE_DESC		"Xilinx XDMA Reference Driver"
+#define DRV_MODULE_NAME		"hermes"
+#define DRV_MODULE_DESC		"Hermes driver"
 
 static char version[] =
 	DRV_MODULE_DESC " " DRV_MODULE_NAME " v" DRV_MODULE_VERSION "\n";
 
 MODULE_AUTHOR("Xilinx, Inc.");
+MODULE_AUTHOR("Eideticom Inc.");
 MODULE_DESCRIPTION(DRV_MODULE_DESC);
 MODULE_VERSION(DRV_MODULE_VERSION);
 MODULE_LICENSE("GPL v2");
@@ -50,71 +51,7 @@ MODULE_LICENSE("GPL v2");
 static int xpdev_cnt;
 
 static const struct pci_device_id pci_ids[] = {
-	{ PCI_DEVICE(0x10ee, 0x9048), },
-	{ PCI_DEVICE(0x10ee, 0x9044), },
-	{ PCI_DEVICE(0x10ee, 0x9042), },
-	{ PCI_DEVICE(0x10ee, 0x9041), },
-	{ PCI_DEVICE(0x10ee, 0x903f), },
-	{ PCI_DEVICE(0x10ee, 0x9038), },
-	{ PCI_DEVICE(0x10ee, 0x9028), },
-	{ PCI_DEVICE(0x10ee, 0x9018), },
-	{ PCI_DEVICE(0x10ee, 0x9034), },
-	{ PCI_DEVICE(0x10ee, 0x9024), },
-	{ PCI_DEVICE(0x10ee, 0x9014), },
-	{ PCI_DEVICE(0x10ee, 0x9032), },
-	{ PCI_DEVICE(0x10ee, 0x9022), },
-	{ PCI_DEVICE(0x10ee, 0x9012), },
-	{ PCI_DEVICE(0x10ee, 0x9031), },
-	{ PCI_DEVICE(0x10ee, 0x9021), },
-	{ PCI_DEVICE(0x10ee, 0x9011), },
-
-	{ PCI_DEVICE(0x10ee, 0x8011), },
-	{ PCI_DEVICE(0x10ee, 0x8012), },
-	{ PCI_DEVICE(0x10ee, 0x8014), },
-	{ PCI_DEVICE(0x10ee, 0x8018), },
-	{ PCI_DEVICE(0x10ee, 0x8021), },
-	{ PCI_DEVICE(0x10ee, 0x8022), },
-	{ PCI_DEVICE(0x10ee, 0x8024), },
-	{ PCI_DEVICE(0x10ee, 0x8028), },
-	{ PCI_DEVICE(0x10ee, 0x8031), },
-	{ PCI_DEVICE(0x10ee, 0x8032), },
-	{ PCI_DEVICE(0x10ee, 0x8034), },
-	{ PCI_DEVICE(0x10ee, 0x8038), },
-
-	{ PCI_DEVICE(0x10ee, 0x7011), },
-	{ PCI_DEVICE(0x10ee, 0x7012), },
-	{ PCI_DEVICE(0x10ee, 0x7014), },
-	{ PCI_DEVICE(0x10ee, 0x7018), },
-	{ PCI_DEVICE(0x10ee, 0x7021), },
-	{ PCI_DEVICE(0x10ee, 0x7022), },
-	{ PCI_DEVICE(0x10ee, 0x7024), },
-	{ PCI_DEVICE(0x10ee, 0x7028), },
-	{ PCI_DEVICE(0x10ee, 0x7031), },
-	{ PCI_DEVICE(0x10ee, 0x7032), },
-	{ PCI_DEVICE(0x10ee, 0x7034), },
-	{ PCI_DEVICE(0x10ee, 0x7038), },
-
-	{ PCI_DEVICE(0x10ee, 0x6828), },
-	{ PCI_DEVICE(0x10ee, 0x6830), },
-	{ PCI_DEVICE(0x10ee, 0x6928), },
-	{ PCI_DEVICE(0x10ee, 0x6930), },
-	{ PCI_DEVICE(0x10ee, 0x6A28), },
-	{ PCI_DEVICE(0x10ee, 0x6A30), },
-	{ PCI_DEVICE(0x10ee, 0x6D30), },
-
-	{ PCI_DEVICE(0x10ee, 0x4808), },
-	{ PCI_DEVICE(0x10ee, 0x4828), },
-	{ PCI_DEVICE(0x10ee, 0x4908), },
-	{ PCI_DEVICE(0x10ee, 0x4A28), },
-	{ PCI_DEVICE(0x10ee, 0x4B28), },
-
-	{ PCI_DEVICE(0x10ee, 0x2808), },
-	{ PCI_DEVICE(0x1d0f, 0xf000), },
-	{ PCI_DEVICE(0x1d0f, 0xf001), },
-
-#ifdef INTERNAL_TESTING
-	{ PCI_DEVICE(0x1d0f, 0x1042), 0},
-#endif
+	{ PCI_DEVICE(0x1de5, 0x3000), },
 	{0,}
 };
 MODULE_DEVICE_TABLE(pci, pci_ids);
@@ -334,7 +271,7 @@ static struct pci_driver pci_driver = {
 	.err_handler = &xdma_err_handler,
 };
 
-static int __init xdma_mod_init(void)
+static int __init hermes_mod_init(void)
 {
 	int rv;
 
@@ -352,7 +289,7 @@ static int __init xdma_mod_init(void)
 	return pci_register_driver(&pci_driver);
 }
 
-static void __exit xdma_mod_exit(void)
+static void __exit hermes_mod_exit(void)
 {
 	/* unregister this driver from the PCI bus driver */
 	dbg_init("pci_unregister_driver.\n");
@@ -360,5 +297,5 @@ static void __exit xdma_mod_exit(void)
 	xdma_cdev_cleanup();
 }
 
-module_init(xdma_mod_init);
-module_exit(xdma_mod_exit);
+module_init(hermes_mod_init);
+module_exit(hermes_mod_exit);
